+++ CHANGELOG.md
@@ -6,6 +6,8 @@ F: fix bug
 ## 位置版本 (2017-03-23)
 * `F` 修复 alert 配置的bug 。
 * `F` calendar 组件，icon 替换成原生的 scss 写的 icon
+* `FA` Captcha 组件，conf example 配置错误。增加未配置的警告。
+* `F` cropper 修复背景 image 。
 
 ## 1.3.2 (2017-03-22)